feat(design-system): enhance color tokens with RGB values and improve dark mode selectors
- Add RGB color variables for accent, success, and error colors in both light and dark modes to support alpha compositing in rgba() functions
- Extract parseColorToRGB helper function to test-utils for reusability across test files
- Refactor dark mode selectors to use :where() pseudo-class for improved specificity handling and consistency
- Update component styles to use CSS custom properties for RGB values instead of hardcoded hex values
- Improve select input styling with mask-image for theme-aware dropdown arrow coloring
- Add transition cleanup in ThemeToggle to prevent style transition side effects
- Enhance theme validation in getStoredTheme to use VALID_THEMES array for consistency
