+++ src/styles/tokens/colors.css
@@ -51,6 +51,11 @@
   --color-accent-default: #3b82f6;
   --color-accent-dark: #2563eb;
   --color-accent: var(--color-accent-default);
+  
+  /* RGB values for alpha compositing */
+  --color-accent-rgb: 59, 130, 246;
+  --color-success-rgb: 34, 197, 94;
+  --color-error-rgb: 239, 68, 68;
 
   /* ============================================
      SEMANTIC COLORS
@@ -104,6 +109,11 @@
   --color-accent-default: #60a5fa;
   --color-accent-dark: #3b82f6;
   --color-accent: var(--color-accent-default);
+  
+  /* RGB values for alpha compositing (dark mode) */
+  --color-accent-rgb: 96, 165, 250;
+  --color-success-rgb: 74, 222, 128;
+  --color-error-rgb: 248, 113, 113;
 
   /* Dark Mode - Semantic Colors (adjusted for dark backgrounds) */
   --color-success: #4ade80;
