+++ src/styles/tokens/components.css
@@ -191,13 +191,15 @@
 }
 
 /* Dark mode adjustments for disabled states */
-.dark .btn-primary:disabled {
+:where(.dark, .dark *) .btn-primary:disabled,
+:where(.dark, .dark *).btn-primary:disabled {
   background-color: var(--color-neutral-700);
   border-color: var(--color-neutral-700);
   color: var(--color-neutral-500);
 }
 
-.dark .btn-secondary:disabled {
+:where(.dark, .dark *) .btn-secondary:disabled,
+:where(.dark, .dark *).btn-secondary:disabled {
   border-color: var(--color-neutral-700);
   color: var(--color-neutral-600);
 }
@@ -236,7 +238,7 @@
 .input:focus {
   outline: none;
   border-color: var(--color-accent);
-  box-shadow: 0 0 0 3px rgba(59, 130, 246, 0.15);
+  box-shadow: 0 0 0 3px rgba(var(--color-accent-rgb), 0.15);
 }
 
 .input:disabled {
@@ -252,7 +254,7 @@
 
 .input-error:focus {
   border-color: var(--color-error);
-  box-shadow: 0 0 0 3px rgba(239, 68, 68, 0.15);
+  box-shadow: 0 0 0 3px rgba(var(--color-error-rgb), 0.15);
 }
 
 /* Input success state */
@@ -262,7 +264,7 @@
 
 .input-success:focus {
   border-color: var(--color-success);
-  box-shadow: 0 0 0 3px rgba(34, 197, 94, 0.15);
+  box-shadow: 0 0 0 3px rgba(var(--color-success-rgb), 0.15);
 }
 
 /* Textarea */
@@ -289,7 +291,7 @@
 .textarea:focus {
   outline: none;
   border-color: var(--color-accent);
-  box-shadow: 0 0 0 3px rgba(59, 130, 246, 0.15);
+  box-shadow: 0 0 0 3px rgba(var(--color-accent-rgb), 0.15);
 }
 
 /* Select input */
@@ -302,10 +304,6 @@
   line-height: 1.5;
   color: var(--color-text-primary);
   background-color: var(--color-bg-primary);
-  background-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' fill='none' viewBox='0 0 20 20'%3e%3cpath stroke='%236b7280' stroke-linecap='round' stroke-linejoin='round' stroke-width='1.5' d='M6 8l4 4 4-4'/%3e%3c/svg%3e");
-  background-position: right var(--spacing-2) center;
-  background-repeat: no-repeat;
-  background-size: 1.5em 1.5em;
   border: var(--border-width-default) solid var(--color-border-default);
   border-radius: var(--radius-default);
   appearance: none;
@@ -314,10 +312,35 @@
               box-shadow var(--transition-fast);
 }
 
+/* Select dropdown arrow using mask-image for theme-aware coloring */
+.select-arrow {
+  position: relative;
+}
+
+.select-arrow::after {
+  content: '';
+  position: absolute;
+  right: var(--spacing-2);
+  top: 50%;
+  transform: translateY(-50%);
+  width: 1.5em;
+  height: 1.5em;
+  background-color: var(--color-text-secondary);
+  mask-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' fill='none' viewBox='0 0 20 20'%3e%3cpath stroke='currentColor' stroke-linecap='round' stroke-linejoin='round' stroke-width='1.5' d='M6 8l4 4 4-4'/%3e%3c/svg%3e");
+  mask-repeat: no-repeat;
+  mask-position: center;
+  mask-size: contain;
+  -webkit-mask-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' fill='none' viewBox='0 0 20 20'%3e%3cpath stroke='currentColor' stroke-linecap='round' stroke-linejoin='round' stroke-width='1.5' d='M6 8l4 4 4-4'/%3e%3c/svg%3e");
+  -webkit-mask-repeat: no-repeat;
+  -webkit-mask-position: center;
+  -webkit-mask-size: contain;
+  pointer-events: none;
+}
+
 .select:focus {
   outline: none;
   border-color: var(--color-accent);
-  box-shadow: 0 0 0 3px rgba(59, 130, 246, 0.15);
+  box-shadow: 0 0 0 3px rgba(var(--color-accent-rgb), 0.15);
 }
 
 /* Label */
@@ -341,18 +364,23 @@
 }
 
 /* Dark mode focus ring adjustments */
-.dark .input:focus,
-.dark .textarea:focus,
-.dark .select:focus {
-  box-shadow: 0 0 0 3px rgba(96, 165, 250, 0.25);
+:where(.dark, .dark *) .input:focus,
+:where(.dark, .dark *).input:focus,
+:where(.dark, .dark *) .textarea:focus,
+:where(.dark, .dark *).textarea:focus,
+:where(.dark, .dark *) .select:focus,
+:where(.dark, .dark *).select:focus {
+  box-shadow: 0 0 0 3px rgba(var(--color-accent-rgb), 0.25);
 }
 
-.dark .input-error:focus {
-  box-shadow: 0 0 0 3px rgba(248, 113, 113, 0.25);
+:where(.dark, .dark *) .input-error:focus,
+:where(.dark, .dark *).input-error:focus {
+  box-shadow: 0 0 0 3px rgba(var(--color-error-rgb), 0.25);
 }
 
-.dark .input-success:focus {
-  box-shadow: 0 0 0 3px rgba(74, 222, 128, 0.25);
+:where(.dark, .dark *) .input-success:focus,
+:where(.dark, .dark *).input-success:focus {
+  box-shadow: 0 0 0 3px rgba(var(--color-success-rgb), 0.25);
 }
