+++ src/styles/tokens/spacing.test.ts
@@ -16,15 +16,27 @@ const spacingCSS = fs.readFileSync(spacingPath, 'utf-8');
 
 /**
  * Extract CSS custom property values from the CSS content
+ * Uses brace-matching to correctly handle nested CSS blocks
  */
 function extractCSSVariables(css: string, selector: string = ':root'): Map<string, string> {
   const variables = new Map<string, string>();
+  const escapedSelector = selector.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
+  const selectorRegex = new RegExp(`${escapedSelector}\\s*\\{`, 'g');
   
-  const selectorRegex = new RegExp(`${selector.replace('.', '\\.')}\\s*\\{([^}]+)\\}`, 'g');
-  const matches = css.matchAll(selectorRegex);
-  
-  for (const match of matches) {
-    const block = match[1];
+  let match;
+  while ((match = selectorRegex.exec(css)) !== null) {
+    // Find matching closing brace with proper nesting
+    let depth = 1;
+    const start = match.index + match[0].length;
+    let i = start;
+    
+    while (i < css.length && depth > 0) {
+      if (css[i] === '{') depth++;
+      else if (css[i] === '}') depth--;
+      i++;
+    }
+    
+    const block = css.slice(start, i - 1);
     const varRegex = /--([\w-]+):\s*([^;]+);/g;
     let varMatch;
     while ((varMatch = varRegex.exec(block)) !== null) {
