+++ src/styles/tokens/typography.test.ts
@@ -9,6 +9,7 @@ import * as fc from 'fast-check';
 import * as fs from 'fs';
 import * as path from 'path';
 import { describe, expect, it } from 'vitest';
+import { getContrastRatio } from '../../test-utils/color-helpers';
 
 // Read and parse the colors.css file for contrast testing
 const colorsPath = path.join(__dirname, 'colors.css');
@@ -33,62 +34,6 @@ function extractCSSVariables(css: string, selector: string = ':root'): Map<strin
   }
   
   return variables;
-}
-
-/**
- * Parse a color value to RGB components
- * Supports hex colors (#RGB, #RRGGBB)
- */
-function parseColorToRGB(color: string): { r: number; g: number; b: number } | null {
-  const hexMatch = color.match(/^#([0-9a-fA-F]{3}|[0-9a-fA-F]{6})$/);
-  if (hexMatch) {
-    const hex = hexMatch[1];
-    if (hex.length === 3) {
-      return {
-        r: parseInt(hex[0] + hex[0], 16),
-        g: parseInt(hex[1] + hex[1], 16),
-        b: parseInt(hex[2] + hex[2], 16),
-      };
-    } else {
-      return {
-        r: parseInt(hex.slice(0, 2), 16),
-        g: parseInt(hex.slice(2, 4), 16),
-        b: parseInt(hex.slice(4, 6), 16),
-      };
-    }
-  }
-  return null;
-}
-
-/**
- * Calculate relative luminance according to WCAG 2.1
- * https://www.w3.org/WAI/GL/wiki/Relative_luminance
- */
-function getRelativeLuminance(rgb: { r: number; g: number; b: number }): number {
-  const sRGB = [rgb.r / 255, rgb.g / 255, rgb.b / 255];
-  const [r, g, b] = sRGB.map((c) => {
-    return c <= 0.03928 ? c / 12.92 : Math.pow((c + 0.055) / 1.055, 2.4);
-  });
-  return 0.2126 * r + 0.7152 * g + 0.0722 * b;
-}
-
-/**
- * Calculate contrast ratio between two colors according to WCAG 2.1
- * https://www.w3.org/WAI/GL/wiki/Contrast_ratio
- */
-function getContrastRatio(color1: string, color2: string): number | null {
-  const rgb1 = parseColorToRGB(color1);
-  const rgb2 = parseColorToRGB(color2);
-  
-  if (!rgb1 || !rgb2) return null;
-  
-  const l1 = getRelativeLuminance(rgb1);
-  const l2 = getRelativeLuminance(rgb2);
-  
-  const lighter = Math.max(l1, l2);
-  const darker = Math.min(l1, l2);
-  
-  return (lighter + 0.05) / (darker + 0.05);
 }
 
 // Extract variables from both :root and .dark selectors
